Removed hover indicating border
This is just a personal preference, but the border used to indicate hovering over applications in the dock was not aesthetically well designed, nor was it necessary. When hovering over an application a label is shown anyway, so there is no need to have a misplaced gray border around the app icon as well. There isn't any hover effect in the real macOS Sierra either, if app icon enlargement is turned off. An option to enlarge app icons upon hovering would be nice, but that is a whole different case.